Always use prebuilt_any in prebuilt etc gen
We now use the Dsts property to manage the
target destinations of the copy rule.

Before there was special care put in place to not
use Dsts in case src and dst filename matched, but
this broke prebuilt_any usage in these cases.

We can't easily switch to prebuilt_any in those cases
unless we also remove support for prebuilt_root as
this specfic module doesn't support have the Dsts
property to avoid installing files in arbitrary locations,
so we're "forced" to do the whole migration as a whole.

Also add a few test cases to ensure prebuilt_any works
as expected (the latter test woudln't pass before)

Test: m --no-skip-soong-tests
Change-Id: I0c9985a6174090ce3ae6bf3201cca090377f5967
Signed-off-by: tejas101k <tejassingh649@rediffmail.com>